doWrite 작성자 정보 저장, doModify, doDelete 권한체크

+++ src/main/java/com/koreaIT/demo/controller/UsrArticleController.java
@@ -11,6 +11,8 @@ import com.koreaIT.demo.util.Util;
 import com.koreaIT.demo.vo.Article;
 import com.koreaIT.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 	
@@ -22,7 +24,11 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 		
 		if (Util.empty(title)) {
 			return ResultData.from("F-1", "제목을 입력해주세요");
@@ -32,7 +38,7 @@ public class UsrArticleController {
 			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 		
-		articleService.writeArticle(title, body);
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
 		
 		int id = articleService.getLastInsertId();
 		
@@ -67,12 +73,20 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession session, int id, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+		}
+		
+		if (article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
 		}
 		
 		articleService.modifyArticle(id, title, body);
@@ -82,12 +96,20 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if (article == null) {
 			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
+		}
+		
+		if (article.getMemberId() != (int) session.getAttribute("loginedMemberId")) {
+			return ResultData.from("F-A", "해당 게시물에 대한 권한이 없습니다");
 		}
 		
 		articleService.deleteArticle(id);
